Duplicates Refactor Test added
