+++ src/main/java/employee_csv_project/controller/csv_controller/duplication_handler/CheckForDuplicates.java
@@ -42,4 +42,8 @@ public class CheckForDuplicates {
     public static ArrayList<EmployeeDTO> getListOfDuplicatedEmployees() {
         return listOfDuplicatedEmployees;
     }
+
+    public static void setListOfNonDuplicatedEmployees(ArrayList<EmployeeDTO> listOfNonDuplicatedEmployees) {
+        CheckForDuplicates.listOfNonDuplicatedEmployees = listOfNonDuplicatedEmployees;
+    }
 }
