+++ src/test/java/employee_csv_project/controller/csv_controller/duplication_handler/CheckForDuplicatesTest.java
@@ -1,4 +1,4 @@
-package employee_csv_project.controller;
+package employee_csv_project.controller.csv_controller.duplication_handler;
 
 import employee_csv_project.controller.csv_controller.duplication_handler.CheckForDuplicates;
 import employee_csv_project.model.EmployeeDTO;
